* utility.c (printdata): Cast argument to isprint() to int
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13382 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/telnet/telnetd/utility.c
@@ -1212,7 +1212,7 @@ printdata(tag, ptr, cnt)
 		for (i = 0; i < 20 && cnt; i++) {
 			sprintf(nfrontp, "%02x", *ptr);
 			nfrontp += strlen(nfrontp); 
-			if (isprint(*ptr)) {
+			if (isprint((int) *ptr)) {
 				xbuf[i] = *ptr;
 			} else {
 				xbuf[i] = '.';
